refactor: Adding ArchiModel interface
